requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Django==2.2.7
djangorestframework==3.10.3
minepy==1.2.4
numpy==1.17.4
Pillow==6.2.1
psycopg2-binary==2.8.4
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
